Add enhanced configuration and make Google Webmaster Meta Tag optional
A google-site-verification meta tag for Google Webmaster Tools is hard
coded into the code.  this content key  may not be expected behavior
for other developers who are cloning or using the site.

To fix, a popular configuration library called rc was added.  A basic
starter configuration file for the project was created and the
google-site-verification content option was moved from the hard coded
location in app.js to the config file.
EJS was also adjusted to optionally display this tag as needed.

If desired, the configuration enhancement can also be used for other
configuration requirements.