Refactoring to fix save on page reload.
There was a bug that prevented you from saving to GitHub when you
reloaded Dillinger. The reason was because branch was not being passed
down the chain. Because of the way we pass in private files, we have to
instead inject the branch property into one of the JSON objects that
gets passed back.

Also renamed *currentOrg to *currentOwner for some consistency, and
org to owner, repoName to repo.