Add pagination of github repos when importing
When a list of github repos that a user can possibly import from gets to
large for mobile users to fit their screen it will become impossible for
them to scroll down the modal window to find the one they want if it's
not in the first ~15 repos.

By adding pagination much like how the google drive documents are paginated we
eliminate that problem.

Fixes #508